Pulling final

+++ WPF App/WPF App/View/View.cs
@@ -137,7 +137,7 @@ namespace WPF_App.View
             ImpovementsLevelsLabels[index].Content = updatedLevel.ToString();
         }
 
-        public void SetButtonsTexts(params int[] prices)
+        public void SetButtonsTexts(params double[] prices)
         {
             for (int i = 0; i < buttons.Count; i++)
             {
@@ -145,7 +145,7 @@ namespace WPF_App.View
             }
         }
 
-        public void SetButtonText(int price, int index)
+        public void SetButtonText(double price, int index)
         {
             
             buttons[index].Content = ShortNumbersMaker(price);
